pyser_ndcl_2/server/requirements.txt
grpcio
grpcio-tools
google_pasta